% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI_predictions.R
\name{MI_predictions}
\alias{MI_predictions}
\title{Make predictions from multiple imputation models}
\usage{
MI_predictions(
  data,
  outcome_var,
  variable_x,
  subgroups = NULL,
  covariables = NULL,
  rcs_terms = NULL,
  imp_col = ".imp",
  model_type = "lm",
  followup_offset = "No",
  followup_col = NULL,
  time_col = NULL,
  event_col = NULL,
  random_intercept = "No",
  random_intercept_var = NULL,
  plot_options = list(title = "Predicted Values with 95\% Confidence Intervals", x_label
    = NULL, y_label = "Predicted Value", line_color = "blue", ribbon_fill = "blue",
    ribbon_alpha = 0.2),
  prediction_range = c(0.01, 0.99),
  n_points = 100
)
}
\arguments{
\item{data}{A data frame containing the imputed dataset.}

\item{outcome_var}{The dependent variable for the model.}

\item{variable_x}{The continuous predictor variable to be modeled.}

\item{subgroups}{Optional variable for stratification (default is NULL).}

\item{covariables}{Optional vector of covariates to adjust for.}

\item{rcs_terms}{Optional list of restricted cubic spline specifications (default is NULL).}

\item{imp_col}{The column name indicating imputation indices (default is ".imp").}

\item{model_type}{The model type: "nb" for Negative Binomial, "poisson" for Poisson, "cox" for Cox, "bin"/"logistic" for Logistic, or "lm" for Linear.}

\item{followup_offset}{Whether to include offset for follow-up duration ("Yes" or "No").}

\item{followup_col}{The column representing follow-up duration (required if followup_offset = "Yes").}

\item{time_col}{The time variable for Cox regression (only required if model_type = "cox").}

\item{event_col}{The event variable for Cox regression (only required if model_type = "cox").}

\item{random_intercept}{Whether to include a random intercept in the model ("Yes" or "No").}

\item{random_intercept_var}{The grouping variable for the random intercept (required if random_intercept = "Yes").}

\item{plot_options}{List of options for plot customization.}

\item{prediction_range}{Range for predictions as quantiles (default is c(0.01, 0.99)).}

\item{n_points}{Number of points to use in the prediction sequence (default is 100).}
}
\value{
A list containing predictions, plots, and model information for each subgroup.
}
\description{
This function makes prediction based on the estimation of the coefficient and standard error obtained with the Rubin rule.
When subgroups are specified, separate models are fit for each subgroup.
}
